Zip Addresses in Tract Addresses?

To use this function, the HUD_KEY system environment variable must be set first: use hud_set_key("the_key") to do this.

Given zip code(s) and tract(s), determine if they overlap using the crosswalk files. Overlap will be described if any residential, business, other, or total addresses reside in both.

This means that it is possible that certain geoids(lhs) are not counted even though their boundaries intersect the queried geoids(rhs). This is likely because addresses do not lie in their intersecting region.

Infix operators default to most recent year and quarter of which the crosswalk files are available. For modifying the year and quarter parameter take a look at their non-infix versions.

Usage

zip %z_in_trt% tract

Arguments

zip

The zip(s) to determine overlap with tract(s)

tract

The tract(s) to determine overlap with zip(s)

Value

If zip(s) exist in the tract(s) specified, then TRUE is returned.

Examples

if (FALSE) {

71052 %z_in_trt% 22031950600

}