hud_cw_cbsa_zip

This function queries the Crosswalks API provided by US Department of Housing and Urban Development. This returns the crosswalk for cbsa to zip.

hud_cw_cbsa_zip(
  cbsa,
  year = format(Sys.Date() - 365, "%Y"),
  quarter = 1,
  minimal = FALSE,
  key = Sys.getenv("HUD_KEY")
)

Arguments

cbsa

5 digit CBSA code for Micropolitan and Metropolitan Areas Eg: 10380 for type 8

year

Gets the year that this data was recorded. Can specify multiple years. Default is the previous year.

quarter

Gets the quarter of the year that this data was recorded. Defaults to the first quarter of the year.

minimal

Return just the crosswalked GEOIDs if true. Otherwise, return all fields. This does not remove duplicates.

key

The API key for this user. You must go to HUD and sign up for an account and request for an API key.

Value

This function returns a dataframe containing CROSSWALK data for a particular GEOID. These measurements include res-ratio, bus-ratio, oth-ratio, tot-ratio. For more details on these measurements, visit https://www.huduser.gov/portal/dataset/uspszip-api.html

Examples

if (FALSE) {
library(hudr)

hud_cw_cbsa_zip(cbsa = 10140, year = c('2017'), quarter = c('1'))

hud_cw_cbsa_zip(cbsa = '10140', year = c('2010'), quarter = c('1'),
   minimal = TRUE)
}