Get Most Recently Released Year: Fair Markets Rent

Pings the Fair Markets Rent API provided by (US Department of Housing and Urban Development) HUD User to determine the most recently released files. This will only ping for the last two years. Will return years for state, county, and metroarea resolutions.

Usage

hud_rec_fmr_yr(key = Sys.getenv("HUD_KEY"))

Arguments

key

The key obtained from HUD (US Department of Housing and Urban Development) USER website.

Value

The most recent year for the fmr files for state, county, and metroarea queries.

See also

Examples

if (FALSE) {

hud_rec_fmr_yr()

}