z_in_ctysb

Given a zip code and a countysub, determine if they overlap using the crosswalk files. Overlap will be described if any residential, business, other, or total addresses reside in both.

z_in_ctysb(zip, countysub, year, quarter)

Arguments

zip

The zip to determine overlap with countysub

countysub

The countysub to determine overlap with zip

year

The year of the crosswalk files.

quarter

The quarter of the crosswalk files.

Value

If zip(s) exist in the countysub(s) specified, then TRUE is returned.

See also

Examples

if (FALSE) {

z_in_ctysb(zip = 35213, countysub = 0107390324, year = 2019, quarter = 2)

}