crosswalk

Using the crosswalk files, crosswalk an entire dataset. Currently supported crosswalks:

zip-tract

zip-county

zip-cbsa

zip-cbsadiv (Available 4th Quarter 2017 onwards)

zip-cd

tract-zip

county-zip

cbsa-zip

cbsadiv-zip (Available 4th Quarter 2017 onwards)

cd-zip

zip-countysub (Available 2nd Quarter 2018 onwards)

countysub-zip (Available 2nd Quarter 2018 onwards)

crosswalk(
  data,
  geoid,
  geoid_col,
  cw_geoid,
  cw_geoid_col = NA,
  method = NA,
  year = format(Sys.Date() - 365, "%Y"),
  quarter = 1,
  key = Sys.getenv("HUD_KEY")
)

Arguments

data

A dataset with rows describing measurements at a zip, county, countysub, cd, tract, cbsa, or cbsadiv geographic id.

geoid

The current geoid that the dataset is described in: must be zip, county, countysub, cd, tract, cbsa, or cbsadiv geographic id.

geoid_col

The column containing the geographic identifier; must be zip, county, countysub, cd, tract, cbsa, or cbsadiv geographic id. Supply either the name of the column or the index. All elements in this column must be numbers only at the proper length. For example, zip codes must be 5 digit numbers.

cw_geoid

The geoid to crosswalk the dataset to.

cw_geoid_col

The columns in the dataset to distribute according to method ratio. If method is empty, no allocation method will be applied -- the crosswalk file will just be merged to the dataset. All elements in these columns must be numbers only.

method

The allocation method to use: residential, business, other, or total. If method is empty, no allocation method will be applied -- the crosswalk file will just be merged to the dataset.

year

The year measurement was taken.

quarter

The quarter of year measurement was taken.

key

The key obtain from HUD USER website.

Value

A dataframe containing the crosswalked dataset.

Examples

if (FALSE) {
library(rhud)

sample <- data.frame(population = c(42134, 12413, 13132),
                     county = c(24047, 24045, 24043))

crosswalk(data = sample, geoid = "county", geoid_col = "county",
          cw_geoid = "zip")

crosswalk(data = sample, geoid = "county", geoid_col = "county",
          cw_geoid = "zip", cw_geoid_col = "population", method = "res")

crosswalk(data = sample, geoid = "county", geoid_col = "county",
          cw_geoid = "zip", cw_geoid_col = "population", method = "bus")

crosswalk(data = sample, geoid = "county", geoid_col = "county",
          cw_geoid = "zip", cw_geoid_col = "population", method = "bus",
          year = 2018, quarter = 1)

}